+++ assets/js/code_sharing.js
@@ -4,9 +4,11 @@ var localtunnel = require('localtunnel');
 const io_client = require('socket.io-client');
 var socket;
 var clients = [];
+var remoteFile;
+var remote_mode = false;
 
-function shareCode(){
-    if(isCurrentFileSaved()){
+function shareCode(file, project_path){
+    // if(isCurrentFileSaved()){
         $('#code-sharing-box').html('<h5>Waiting...</h5>');
         server.listen(4000, function(){
             // console.log('listening at 4000');
@@ -31,17 +33,38 @@ function shareCode(){
         // console.log(file);
         io.on('connection', function (socket) {
             clients.push(socket);
-          socket.emit('file', { id: file.id, name:file.name, remote_path:file.path, data:file.editor.getValue() });
-          socket.on('commit-changes', function (data) {
-            // console.log('hello');
-            if(file.id == data.id && file.path == data.path){
-                file.editor.setValue(data.data);
-                editor.setValue(data.data);
+            if(file){
+                socket.emit('file', { id: file.id, name:file.name, remote_path:file.path, data:file.editor.getValue() });
+            }else{
+                socket.emit('project', {path: project_path, name : path.basename(project_path)});
             }
-            files['#' + data.id].editor.setValue(data.data);
-          });
+            socket.on('commit-changes', function (data) {
+            // console.log('hello');
+                if(file.id == data.id && file.path == data.path){
+                    file.editor.setValue(data.data);
+                    editor.setValue(data.data);
+                }
+                files['#' + data.id].editor.setValue(data.data);
+            });
+            socket.on('getRemoteDirectoryForSpecificDirpath', function(folder_path){
+                // console.log(folder_path);
+                let dir_structure = createDirectoryForSpecificDirpath(folder_path, true);
+                // console.log(dir_structure);
+                this.emit('directoryStructureForSpecificRemoteDirpath', dir_structure);
+            });
+            socket.on('getContentOfRemoteFile', function(filepath){
+                console.log('filepath')
+                let temp_file = openFileFromSidebar(filepath, true);
+                console.log(temp_file);
+                this.emit('openFile', temp_file);
+            });
+
+            socket.on('saveRemoteFile', function(data, remote_path){
+                let status = saveFile(data, remote_path, true);
+                console.log(status);
+            });
         });
-    }
+    // }
 }
 
 function connect(flag){
@@ -61,6 +84,24 @@ function connect(flag){
             $('#code-sharing-box').html('<button class = "btn" id = "commit_changes" onclick="code_sharing.commitChanges()">Commit Changes</button><br><br>');
             $('#code-sharing-box').append('<button class = "btn" id = "close_socket" onclick="code_sharing.closeSocket()">Close Connection</button>');
         });
+
+        socket.on('project', function(data){
+            code_sharing.remote_mode = true;
+            // console.log(data);
+            openFolder(data);
+            // $('#code-sharing-box').html('<button class = "btn" id = "commit_changes" onclick="code_sharing.commitChanges()">Commit Changes</button><br><br>');
+            $('#code-sharing-box').html('<button class = "btn" id = "close_socket" onclick="code_sharing.closeSocket()">Close Connection</button>');
+        });
+
+        socket.on('directoryStructureForSpecificRemoteDirpath', function(dir_structure){
+            console.log(dir_structure);
+            openSpecificDirectory(dir_structure);
+        });
+
+        socket.on('openFile', function(temp_file){
+            openFile(temp_file.data, undefined, temp_file.name, temp_file.filepath);
+        });
+
         socket.on('commit-changes', function(data){
             // console.log('data recieved');
             // console.log(file.remote_path, remoteFile.remote_path);
@@ -94,10 +135,11 @@ function commitChanges(){
 }
 
 function backAllButtons(){
-    $('#code-sharing-box').html('<button class = "btn" id = "share-code" onclick="code_sharing.shareCode()">Share Code</button><br><br><button class = "btn" id = "connect" onclick="code_sharing.connect(false)">Connect</button>');
+    $('#code-sharing-box').html('<button class = "btn" id = "share-code" onclick="code_sharing.shareProject()">Share Project</button><br><br><button class = "btn" id = "share-code" onclick="code_sharing.shareCode()">Share Code</button><br><br><button class = "btn" id = "connect" onclick="code_sharing.connect(false)">Connect</button>');
 }
 
 function closeSocket(){
+    code_sharing.remote_mode = false;
     socket.destroy();
     backAllButtons();
 }
@@ -107,6 +149,30 @@ function disconnect(){
     backAllButtons();
 }
 
+function shareProject(){
+    let project_path = ipc.sendSync('getProjectPath');
+    // console.log(project_path);
+    if(project_path != null){
+        shareCode(false, project_path[0]);
+    }else{
+        code_sharing.remote_mode = false;
+    }
+    // console.log(code_sharing.remote_mode);
+}
+
+function getRemoteDirectoryForSpecificDirpath(folder_path){
+    console.log(folder_path);
+    socket.emit('getRemoteDirectoryForSpecificDirpath', folder_path);
+}
+
+function openRemoteFile(filepath){
+    socket.emit('getContentOfRemoteFile', filepath);
+}
+
+function saveRemoteFile(data, remote_path){
+    socket.emit('saveRemoteFile', data, remote_path);
+}
+
 module.exports = {
 	shareCode : shareCode,
 	connect : connect,
@@ -114,5 +180,8 @@ module.exports = {
 	commitChanges : commitChanges,
 	backAllButtons : backAllButtons,
 	closeSocket : closeSocket,
-	disconnect : disconnect
+	disconnect : disconnect,
+    shareProject : shareProject,
+    getRemoteDirectoryForSpecificDirpath,
+    openRemoteFile : openRemoteFile
 }