+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Add: Colon remember you last working project
 - Fixed: Auto close when all tabs are closed
 - Fixed: Editor will auto focus when you create a new file
+- Fixed: Can share unsaved file as well
+- Improvement: UI